1. refactor message handlers to make each peer has it's own message handler 2. fix peers connection disconnection freeze issue 3. refresh bloom filter by FalsePositiveRate 4. fix transaction input false positive not working issue

+++ spvwallet/db/stxosdb.go
@@ -4,8 +4,8 @@ import (
 	"database/sql"
 	"sync"
 
-	"github.com/ioeX/ioeX.Utility/common"
-	"github.com/ioeX/ioeX.MainChain/core"
+	"github.com/ioeXNetwork/ioeX.MainChain/core"
+	"github.com/ioeXNetwork/ioeX.Utility/common"
 )
 
 const CreateSTXOsDB = `CREATE TABLE IF NOT EXISTS STXOs(
@@ -45,9 +45,18 @@ func (db *STXOsDB) FromUTXO(outPoint *core.OutPoint, spendTxId *common.Uint256, 
 	sql := `INSERT OR REPLACE INTO STXOs(OutPoint, Value, LockTime, AtHeight, ScriptHash, SpendHash, SpendHeight)
 			SELECT UTXOs.OutPoint, UTXOs.Value, UTXOs.LockTime, UTXOs.AtHeight, UTXOs.ScriptHash, ?, ? FROM UTXOs
 			WHERE OutPoint=?`
-	_, err = tx.Exec(sql, spendTxId.Bytes(), spendHeight, outPoint.Bytes())
+	result, err := tx.Exec(sql, spendTxId.Bytes(), spendHeight, outPoint.Bytes())
 	if err != nil {
 		return err
+	}
+
+	rows, err := result.RowsAffected()
+	if err != nil {
+		return nil
+	}
+
+	if rows == 0 {
+		return fmt.Errorf("no rows effected")
 	}
 
 	_, err = tx.Exec("DELETE FROM UTXOs WHERE OutPoint=?", outPoint.Bytes())
